+++ src/compojure/api/meta.clj
@@ -3,7 +3,10 @@
             [compojure.api.common :refer :all]
             [compojure.api.middleware :as mw]
             [compojure.api.exception :as ex]
-            [compojure.core :refer [routes]]
+
+            compojure.core
+            #_clout.core
+
             [plumbing.core :refer :all]
             [plumbing.fnk.impl :as fnk-impl]
             [ring.swagger.common :as rsc]
@@ -52,28 +55,28 @@
 
 (defn fnk-schema [bind]
   (:input-schema
-   (fnk-impl/letk-input-schema-and-body-form
-     nil (with-meta bind {:schema s/Any}) [] nil)))
+    (fnk-impl/letk-input-schema-and-body-form
+      nil (with-meta bind {:schema s/Any}) [] nil)))
 
 (defn body-coercer-middleware [handler coercer responses]
   (fn [request]
     (if-let [{:keys [status] :as response} (handler request)]
       (if-let [schema (:schema (responses status))]
-        (if-let [matcher (:response (mw/get-coercion-matcher-provider request))]
+        (if-let [matcher (:response (mw/coercion-matchers request))]
           (let [coerce (coercer (rsc/value-of schema) matcher)
                 body (coerce (:body response))]
             (if (su/error? body)
               (throw (ex-info "Response validation error"
                               (assoc body :type ::ex/response-validation)))
               (assoc response
-                     ::serializable? true
-                     :body body)))
+                ::serializable? true
+                :body body)))
           response)
         response))))
 
 (defn coerce! [schema key type coercer request]
   (let [value (keywordize-keys (key request))]
-    (if-let [matcher (type (mw/get-coercion-matcher-provider request))]
+    (if-let [matcher (type (mw/coercion-matchers request))]
       (let [coerce (coercer schema matcher)
             result (coerce value)]
         (if (su/error? result)
@@ -276,18 +279,44 @@
   "Dummy letk-macro used in resolving route-docs. not part of normal invokation chain."
   [bindings & body]
   (reduce
-   (fn [cur-body-form [bind-form]]
-     (if (symbol? bind-form)
-       `(let [~bind-form nil] ~cur-body-form)
-       (let [{:keys [map-sym body-form]} (fnk-impl/letk-input-schema-and-body-form
-                                           &env
-                                           (fnk-impl/ensure-schema-metadata &env bind-form)
-                                           []
-                                           cur-body-form)
-             body-form (clojure.walk/prewalk-replace {'plumbing.fnk.schema/safe-get 'clojure.core/get} body-form)]
-         `(let [~map-sym nil] ~body-form))))
-   `(do ~@body)
-   (reverse (partition 2 bindings))))
+    (fn [cur-body-form [bind-form]]
+      (if (symbol? bind-form)
+        `(let [~bind-form nil] ~cur-body-form)
+        (let [{:keys [map-sym body-form]} (fnk-impl/letk-input-schema-and-body-form
+                                            &env
+                                            (fnk-impl/ensure-schema-metadata &env bind-form)
+                                            []
+                                            cur-body-form)
+              body-form (clojure.walk/prewalk-replace {'plumbing.fnk.schema/safe-get 'clojure.core/get} body-form)]
+          `(let [~map-sym nil] ~body-form))))
+    `(do ~@body)
+    (reverse (partition 2 bindings))))
+
+;;
+;; Compojure overrides
+;;
+
+#_(defn- if-context [path route handler]
+  (fn [request]
+    (if-let [params (clout.core/route-matches route request)]
+      (let [uri (:uri request)
+            subpath (:__path-info params)
+            ctx (conj (if-let [ctx (:compojure/context request)] ctx []) path)
+            params (dissoc params :__path-info)]
+        (handler
+          (-> request
+              (#'compojure.core/assoc-route-params (#'compojure.core/decode-route-params params))
+              (assoc :path-info (if (= subpath "") "/" subpath)
+                     :compojure/context ctx
+                     :context (#'compojure.core/remove-suffix uri subpath))))))))
+
+#_(defmacro context [path args & routes]
+  `(#'if-context
+     ~path
+     ~(#'compojure.core/context-route path)
+     (fn [request#]
+       (compojure.core/let-request [~args request#]
+                                   (compojure.core/routing request# ~@routes)))))
 
 ;;
 ;; Api
@@ -326,36 +355,43 @@
                 middlewares
                 parameters
                 body]} (reduce
-                        (fn [acc [k v]]
-                          (restructure-param k v (update-in acc [:parameters] dissoc k)))
-                        (map-of lets letks responses middleware parameters body)
-                        parameters)
-        _ (assert (not middlewares) ":middlewares is deprecated, use :middleware instead.")
+                         (fn [acc [k v]]
+                           (restructure-param k v (update-in acc [:parameters] dissoc k)))
+                         (map-of lets letks responses middleware parameters body)
+                         parameters)
+
+        ;; migration helper
+        _ (assert (not middlewares) ":middlewares is deprecated with 1.0.0, use :middleware instead.")
+
+        ;; response coercion middleware, why not just code?
+        middleware (if responses
+                     (conj middleware `[body-coercer-middleware ~'+compojure-api-coercer+ ~responses])
+                     middleware)
 
         pre-lets [+compojure-api-coercer+ `(memoized-coercer)]
-        wrap (or routes 'do)
+        body-wrap (or routes 'do)
 
-        form `(~wrap ~@body)
+        form `(~body-wrap ~@body)
         form (if (seq letks) `(letk ~letks ~form) form)
         form (if (seq lets) `(let ~lets ~form) form)
         form (if (seq middleware)
                `(let [wrap-mw# (mw/compose-middleware ~middleware)]
                   ((wrap-mw# (fn [~arg] ~form)) ~arg))
                form)
+        form (if (seq pre-lets) `(let ~pre-lets ~form) form)
         form (if routes
                `(compojure.core/context ~path ~arg-with-request ~form)
                (compojure.core/compile-route method path arg-with-request (list form)))
-        form (if responses `(body-coercer-middleware ~form ~+compojure-api-coercer+ ~responses) form)
-        form (if (seq pre-lets) `(let ~pre-lets ~form) form)
 
         ;; for routes, create a separate lookup-function to find the inner routes
         child-form (if routes
-                     (let [form `(~wrap ~@body)
+                     (let [form `(~body-wrap ~@body)
                            form (if (seq letks) `(dummy-letk ~letks ~form) form)
                            form (if (seq lets) `(dummy-let ~lets ~form) form)
-                           form `(fn [~'+compojure-api-request+] ~form)
-                           form (if (seq pre-lets) `(let ~pre-lets ~form) form)]
+                           form `(compojure.core/let-request [~arg ~'+compojure-api-request+] ~form)
+                           form `(fn [~'+compojure-api-request+] ~form)]
                        form))]
-
-    `(let [childs# ~(if routes [`(~child-form {})] nil)]
-       (routes/create ~path-string ~method ~parameters childs# ~form))))
+    (if routes
+      `(let [childs# ~(if routes [`(~child-form {})] nil)]
+         (routes/create ~path-string ~method ~parameters childs# ~form))
+      `(routes/create ~path-string ~method ~parameters nil ~form))))